routewise/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
litellm==1.40.0
scikit-learn==1.4.2
numpy==1.26.4
tiktoken==0.7.0
streamlit==1.35.0
httpx==0.27.0
pydantic==2.7.1
tabulate==0.9.0
